Use either type of quotation marks in string
It may be the case that the user want to use a string that contains
quotation marks. To not limit the user to one particular kind, single or
double qutation mark, a string should be able to be delimeted by either
type.
This will enable the user to use strings like in, for example, Python,
by just choosing the delimiting quotation mark based on the qutation
mark desired inside of the string.