Add tokenizing/parsing of first part of if structure
This commit adds parsing of the general structure of an if, it does not
consider the condition yet. For now, that just get's discarded.

This is also still missing the if body. At the moment only the if keyword,
and the condition in parentheses gets tokenized/parsed. The if-body will
get implemented later.