Fix incorrect tests
The test for the if body was written under the assumption that the
if-body should contain a string.
This is not the desired behaviour though, and therefor incorrect.
The if-body should actually contain any number of tokens, provided they
are in the correct order.
For example a, valid if-statement might be

    'if("some condition that is true"){instruction("do this when true")}

All tokens inside the if-body should follow the exact same rules as if
they were outside of the if, whith the one exception that the if-body
can not contain a name token.

Following from that, there should be a function that parses all the
tokens inside of the if body, that is very similiar to the existing
'parseTokens' function, and get's called everytime a if-token (or any
other upcoming token that can have tokens nested inside of it) is
encountered.